+++ src/main/java/chess/piece/nonsliding/King.java
@@ -1,6 +1,8 @@
-package chess.piece;
+package chess.piece.nonsliding;
 
 import chess.board.Position;
+import chess.piece.Direction;
+import chess.piece.Side;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
